requirements.txt
# Core Trading Dependencies
krakenex==2.2.2
python-kraken-sdk==1.0.0
ccxt==4.1.22
pandas==2.1.3
numpy==1.24.3
ta==0.11.0
python-dotenv==1.0.0

# Web Framework & WebSocket
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
eventlet==0.33.3

# Database
sqlalchemy==2.0.23
sqlite3-python==1.0.0
alembic==1.12.1

# Risk Management & Analysis
scipy==1.11.4
scikit-learn==1.3.2
statsmodels==0.14.0

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk==1.38.0

# Charts & Visualization
plotly==5.18.0
dash==2.14.1

# Security
cryptography==41.0.7
pyjwt==2.8.0
python-jose==3.3.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Async Support
asyncio==3.4.3
aiohttp==3.9.1
websockets==12.0

# Scheduling
apscheduler==3.10.4
schedule==1.2.0

# Utilities
requests==2.31.0
pytz==2023.3
python-dateutil==2.8.2
pyyaml==6.0.1
flask-sqlalchemy
websocket